--- .idea/inspectionProfiles/Project_Default.xml
@@ -1,10 +1,0 @@
-<component name="InspectionProjectProfileManager">
-  <profile version="1.0">
-    <option name="myName" value="Project Default" />
-    <inspection_tool class="DuplicatedCode" enabled="true" level="WEAK WARNING" enabled_by_default="true">
-      <Languages>
-        <language minSize="107" name="Java" />
-      </Languages>
-    </inspection_tool>
-  </profile>
-</component>
