--- .idea/libraries/hynnet_tomcat_catalina.xml
@@ -1,32 +1,0 @@
-<component name="libraryTable">
-  <library name="hynnet.tomcat.catalina" type="repository">
-    <properties maven-id="com.hynnet:tomcat-catalina:8.0.11" />
-    <CLASSES>
-      <root url="jar://$PROJECT_DIR$/lib/tomcat-catalina-8.0.11.jar!/" />
-      <root url="jar://$PROJECT_DIR$/lib/tomcat-servlet-api-8.0.11.jar!/" />
-      <root url="jar://$PROJECT_DIR$/lib/tomcat-jsp-api-8.0.11.jar!/" />
-      <root url="jar://$PROJECT_DIR$/lib/tomcat-el-api-8.0.11.jar!/" />
-      <root url="jar://$PROJECT_DIR$/lib/tomcat-juli-8.0.11.jar!/" />
-      <root url="jar://$PROJECT_DIR$/lib/tomcat-annotations-api-8.0.11.jar!/" />
-      <root url="jar://$PROJECT_DIR$/lib/tomcat-api-8.0.11.jar!/" />
-      <root url="jar://$PROJECT_DIR$/lib/tomcat-jni-8.0.11.jar!/" />
-      <root url="jar://$PROJECT_DIR$/lib/tomcat-coyote-8.0.11.jar!/" />
-      <root url="jar://$PROJECT_DIR$/lib/tomcat-util-8.0.11.jar!/" />
-      <root url="jar://$PROJECT_DIR$/lib/tomcat-util-scan-8.0.11.jar!/" />
-    </CLASSES>
-    <JAVADOC />
-    <SOURCES>
-      <root url="jar://$PROJECT_DIR$/lib/tomcat-catalina-8.0.11-sources.jar!/" />
-      <root url="jar://$PROJECT_DIR$/lib/tomcat-servlet-api-8.0.11-sources.jar!/" />
-      <root url="jar://$PROJECT_DIR$/lib/tomcat-jsp-api-8.0.11-sources.jar!/" />
-      <root url="jar://$PROJECT_DIR$/lib/tomcat-el-api-8.0.11-sources.jar!/" />
-      <root url="jar://$PROJECT_DIR$/lib/tomcat-juli-8.0.11-sources.jar!/" />
-      <root url="jar://$PROJECT_DIR$/lib/tomcat-annotations-api-8.0.11-sources.jar!/" />
-      <root url="jar://$PROJECT_DIR$/lib/tomcat-api-8.0.11-sources.jar!/" />
-      <root url="jar://$PROJECT_DIR$/lib/tomcat-jni-8.0.11-sources.jar!/" />
-      <root url="jar://$PROJECT_DIR$/lib/tomcat-coyote-8.0.11-sources.jar!/" />
-      <root url="jar://$PROJECT_DIR$/lib/tomcat-util-8.0.11-sources.jar!/" />
-      <root url="jar://$PROJECT_DIR$/lib/tomcat-util-scan-8.0.11-sources.jar!/" />
-    </SOURCES>
-  </library>
-</component>
