+++ mes_wood.iml
@@ -21,6 +21,5 @@
     <orderEntry type="inheritedJdk" />
     <orderEntry type="sourceFolder" forTests="false" />
     <orderEntry type="library" name="javax.annotation" level="project" />
-    <orderEntry type="library" name="hynnet.tomcat.catalina" level="project" />
   </component>
 </module>
