--- src/com/mes/controller/account/AccountSearchController.java
@@ -1,15 +1,0 @@
-package com.mes.controller.account;
-
-import javax.servlet.ServletException;
-import javax.servlet.http.HttpServlet;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import java.io.IOException;
-
-public class AccountSearchController extends HttpServlet {
-    @Override
-    protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-
-        super.doPost(req, resp);
-    }
-}
