+++ src/com/mes/dao/MaterialsOrdersDao.java
@@ -43,7 +43,7 @@ public class MaterialsOrdersDao extends DBManager {
             System.out.println("materialOrderDao findAllMaterialsOrders Error");
             return null;
         }finally {
-            closeConnectionAll(rs, ps, conn);
+            closeConnectionAll(conn, ps, rs);
         }
     }
